XWIKI-22677: Better visuals for pagination * Fixed merge conflicts
