+++ xwiki-platform-core/xwiki-platform-livedata/xwiki-platform-livedata-webjar/src/main/vue/src/components/LivedataPagination.test.js
@@ -163,11 +163,13 @@ describe("LivedataPagination.vue", () => {
         },
       },
     });
-    expect(wrapper.find(".pagination-indexes").text()).toContain("livedata.pagination.page");
-    let pageNavs = wrapper.findAll(".pagination-indexes .page-nav");
-    expect(pageNavs.length).toBe(3);
-    expect(pageNavs.at(0).text()).toBe("1");
-    expect(pageNavs.at(1).text()).toBe("2");
-    expect(pageNavs.at(2).text()).toBe("3");
+    let pageNavs = wrapper.findAll('.pagination-indexes .page-nav');
+    expect(pageNavs.length).toBe(3)
+    expect(pageNavs.at(0).text()).toContain("1");
+    expect(pageNavs.at(0).text()).toContain("livedata.pagination.loadPageByNumber");
+    expect(pageNavs.at(1).text()).toContain("2");
+    expect(pageNavs.at(1).text()).toContain("livedata.pagination.loadPageByNumber");
+    expect(pageNavs.at(2).text()).toContain("3");
+    expect(pageNavs.at(2).text()).toContain("livedata.pagination.loadPageByNumber");
   });
 });
